Extracted common functionality to WikiPage class
