+++ prepare_from_template.py
