+++ replace_selected_command.py
